EFGS: Fix isRecent filtering
There may be quite a relevant delay between filtering of too-old keys in `EfgsDownload*Keys` and importing them in `EfgsImportKeys`. This is why we need to do the filtering once again right before the import.
One would think, why we can't filter the keys only once, when importing. We can but that would worsen number of generated batches for import so I think filtering them twice is the best, yet sad, solution.